refactor(bundle): extract lg.go standalone-binary subsystem into pkg/bundle
The trailer format (LGBX/LGB2), payload reader/writer, and resource
collection lived inline in lg.go (package main), which forced its white-box
test to sit at the repo root. Move the subsystem into a new pkg/bundle
package with a small exported API — ReadBundled, BaseBinarySize,
CollectResources, AppendTrailer — and keep the format internals
(parseTrailer, payloadFitsFile, the magic bytes, bundleKind) unexported.
lg.go imports pkg/bundle at its existing call sites; behavior is unchanged.

The bundle-internals tests (payload-fit, trailer parse, resource collection)
move with the code to pkg/bundle/bundle_test.go (package bundle). The
black-box resource e2e tests stay in resources_e2e_test.go for now (they
relocate to test/e2e/ in the next commit).
